Deal with merge conflicts
